♻ [#23] Refactor Charm main access class and transfer functionality to C class
The main responsibility of the Charm class has been shifted to the C class. All methods and properties, including kernel and module access that were previously part of the Charm class, are now in the C class. The Charm class now extends the C class, adopting the so-called "magic magnet" approach to an alternative syntax.