🔧 Allow mixed type inputs in array methods
The `C::Arrays()->get(...)` and `C::Arrays()->has(...)` methods in the Arrays module of the Vivid Kernel have been updated to take any type as an array input which is then cast into an array. This update is documented in the change-log. Additionally, some minor code refactoring was done for better readability and improved implementation.