✨ Add total runtime metric method and refine time measurement
A new method 'getTotalRuntimeMetric' is added in the Performance class to calculate the total runtime using either the output done time or the controller done time. Changes have also been made in the Metric class to improve time measurement. The start and end time can now be set manually and the end time will default to the current time if it's not set.